chore: implement code review suggestions
