+++ bciers/apps/registration/app/components/operations/OperationDataGrid.tsx
@@ -2,34 +2,12 @@
 
 import { useMemo, useState } from "react";
 import DataGrid from "@bciers/components/datagrid/DataGrid";
-import { GridColumnGroupHeaderParams } from "@mui/x-data-grid";
 import OperationsActionCell from "@bciers/components/datagrid/cells/OperationsActionCell";
 import HeaderSearchCell from "@bciers/components/datagrid/cells/HeaderSearchCell";
 import operationColumns from "../datagrid/models/operationColumns";
 import operationGroupColumns from "../datagrid/models/operationGroupColumns";
 import { OperationRow } from "./types";
 import fetchOperationsPageData from "./fetchOperationsPageData";
-
-const OperationSearchCell = ({
-  lastFocusedField,
-  setLastFocusedField,
-}: {
-  lastFocusedField: string | null;
-  setLastFocusedField: (value: string | null) => void;
-}) => {
-  const RenderCell = (params: GridColumnGroupHeaderParams) => {
-    const { groupId, headerName } = params;
-    return (
-      <HeaderSearchCell
-        field={groupId ?? ""}
-        fieldLabel={headerName ?? ""}
-        isFocused={lastFocusedField === groupId}
-        setLastFocusedField={setLastFocusedField}
-      />
-    );
-  };
-  return RenderCell;
-};
 
 const OperationDataGrid = ({
   initialData,
@@ -44,7 +22,7 @@ const OperationDataGrid = ({
   const [lastFocusedField, setLastFocusedField] = useState<string | null>(null);
 
   const SearchCell = useMemo(
-    () => OperationSearchCell({ lastFocusedField, setLastFocusedField }),
+    () => HeaderSearchCell({ lastFocusedField, setLastFocusedField }),
     [lastFocusedField, setLastFocusedField],
   );
