+++ bciers/apps/registration/tests/facilities/mocks.ts
@@ -1,0 +1,9 @@
+const fetchFacilitiesPageData = vi.fn();
+vi.mock(
+  "apps/registration/app/components/facilities/fetchFacilitiesPageData",
+  () => ({
+    default: fetchFacilitiesPageData,
+  }),
+);
+
+export { fetchFacilitiesPageData };
