+++ bciers/libs/shared/testConfig/src/mocks.ts
@@ -30,8 +30,6 @@ const useSearchParams = vi.fn();
 const useSession = vi.fn();
 const auth = vi.fn();
 const fetchOperationsPageData = vi.fn();
-const fetchFacilitiesPageData = vi.fn();
-const createFetchFacilitiesPageData = vi.fn();
 
 export {
   actionHandler,
@@ -43,6 +41,4 @@ export {
   useSearchParams,
   useSession,
   fetchOperationsPageData,
-  fetchFacilitiesPageData,
-  createFetchFacilitiesPageData,
 };
